+++ person.go
@@ -159,7 +159,7 @@ func (h *handler) listPersonPatches(
 			State:   stateVals,
 			OrderBy: order,
 			Size:    defaultPageSize,
-			Skip:    int32((currentPage - 1) * defaultPageSize),
+			Skip:    (currentPage - 1) * defaultPageSize,
 		})
 		if err != nil {
 			return errgo.Wrap(err, "failed to query data")
